# eval: resolveIndex speedup